Introduce generic `properties_treeview.py` and use it in `properties_view.py`
